ground_station/requirements.txt
pyserial
PyQt5
PyQtWebEngine
matplotlib
